fix: 파싱 로직 분리
- chore: 디렉토리 이동
